Move forms SocialContact inputs to containers/forms/
