+++ frontend/containers/forms/website-social/types.ts
@@ -1,6 +1,6 @@
 import { UseFormRegister, RegisterOptions, FieldPath, FormState } from 'react-hook-form';
 
-export type SocialContactInputs = {
+export type WebsiteSocialInputTypes = {
   website: string;
   twitter: string;
   facebook: string;
@@ -8,7 +8,7 @@ export type SocialContactInputs = {
   instagram: string;
 };
 
-export interface InputSocialContactProps<FormValues> {
+export interface WebsiteSocialProps<FormValues> {
   /** React Hook Form's `register` function */
   register: UseFormRegister<FormValues>;
   /** Options for React Hook Form's `register` function */
